Moving mock data & msw logic to tests-related folder
