+++ jest/jest.setup.env.js
@@ -1,9 +1,7 @@
-//const { matchers } = require('@emotion/jest');
 const { loadEnvConfig } = require('@next/env');
 
 const setupEnvForJest = async () => {
   loadEnvConfig('.');
 };
-//const emotionMatchers = matchers;
 
 module.exports = setupEnvForJest
