+++ src/tests-related/mocked-data/mocked-roles.ts
@@ -1,0 +1,6 @@
+export const mockedRoles = [
+  { id: 1, name: 'Developer' },
+  { id: 2, name: 'Devops' },
+  { id: 3, name: 'Squad Owner' },
+  { id: 4, name: 'Chapter Owner' },
+];
