+++ src/tests-related/mocked-data/mocked-skills.ts
@@ -1,0 +1,22 @@
+export const mockedSkills = [
+  {
+    id: 1,
+    name: 'jest',
+  },
+  {
+    id: 2,
+    name: 'prisma',
+  },
+  {
+    id: 3,
+    name: 'nest',
+  },
+  {
+    id: 4,
+    name: 'react',
+  },
+  {
+    id: 5,
+    name: 'Typescript',
+  },
+];
